requirements.txt
flask
eyepop
python-dotenv
pillow
matplotlib
python-quickbooks
intuit-oauth 